:snake: feat: organização de pastas - foi criado uma pasta para cada pagina da pasta pages - foi tirado a pasta css do assets - falta colocar o icon no site e subir para o vercel
