+++ index.html
@@ -2,9 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <!-- <link rel="icon" type="favicon" href="/DigitalCollege/src/assets/favicon/icon_logo.png" /> -->
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
+    <title>DripStore</title>
+    <link rel="shortcut icon" href="/DigitalCollege/src/assets/favicon/icon_logo.png" />
   </head>
   <body>
     <div id="root"></div>
