--- src/pages/CartPage.jsx
@@ -1,17 +1,0 @@
-import Header from "../Components/Header/Header"
-import Footer from "../Components/Footer/Footer"
-import CarouselMenor from "../Components/CarouselMenor/CarouselMenor";
-
-function Cart() {
-
-  return (
-    <>
-      <Header />
-      <h1>Cart</h1>
-      <CarouselMenor />
-      <Footer />
-    </>
-  )
-}
-
-export default Cart
