+++ src/pages/CartPage/CartPage.jsx
@@ -1,0 +1,16 @@
+import Header from "../../Components/Header/Header";
+import Footer from "../../Components/Footer/Footer";
+import CarouselMenor from "../../Components/CarouselMenor/CarouselMenor";
+
+function Cart() {
+  return (
+    <>
+      <Header />
+      <h1>Cart</h1>
+      <CarouselMenor />
+      <Footer />
+    </>
+  );
+}
+
+export default Cart;
