+++ src/pages/ConfirmarCompraPage/ConfirmarCompraPage.jsx
@@ -1,0 +1,14 @@
+import Header from "../../Components/Header/Header";
+import Footer from "../../Components/Footer/Footer";
+
+function ConfirmarCompra() {
+  return (
+    <>
+      <Header />
+      <h1>ConfirmarCompra</h1>
+      <Footer />
+    </>
+  );
+}
+
+export default ConfirmarCompra;
