--- src/pages/CriarFormularioPage.jsx
@@ -1,15 +1,0 @@
-import HeaderMenor from "../Components/HeaderMenor/HeaderMenor"
-import Footer from "../Components/Footer/Footer"
-
-function CriarFormulario() {
-
-  return (
-    <>
-      <HeaderMenor />
-      <h1>CriarFormulario</h1>
-      <Footer />
-    </>
-  )
-}
-
-export default CriarFormulario
