+++ src/pages/CriarFormularioPage/CriarFormularioPage.jsx
@@ -1,0 +1,14 @@
+import HeaderMenor from "../../Components/HeaderMenor/HeaderMenor";
+import Footer from "../../Components/Footer/Footer";
+
+function CriarFormulario() {
+  return (
+    <>
+      <HeaderMenor />
+      <h1>CriarFormulario</h1>
+      <Footer />
+    </>
+  );
+}
+
+export default CriarFormulario;
