--- src/pages/HomePage.css
@@ -1,67 +1,0 @@
-.produto-em-alta-cards {
-          display: flex;
-          justify-content: center;
-          gap: 1.5rem;
-          margin-bottom: 4rem;
-}
-
-.container-produtos-em-alta {
-          display: flex;
-          flex-direction: column;
-          justify-content: space-around;
-          align-items: center;
-}
-
-.produtos-em-alta .topico-section{
-          display: flex;
-          justify-content: space-between;
-          margin-bottom: 1.5rem;
-}
-
-.produtos-em-alta .topico-section h1 {
-          font-size: 1.5rem;
-          font-family: var(--font_primaria);
-          font-weight: var(--font_weight_bold);
-          color: var(--black);
-          letter-spacing: 0.046875rem;
-}
-
-.produtos-em-alta .topico-section h2 {
-          font-size: 1.125rem;
-          font-family: var(--font_primaria);
-          font-weight: var(--font_weight_normal);
-          color: var(--primary);
-          letter-spacing: 0.046875rem;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-}
-
-.produtos-em-alta .topico-section h2 {
-          font-size: 1.125rem;
-          font-family: var(--font_primaria);
-          font-weight: var(--font_weight_normal);
-          color: var(--primary);
-          letter-spacing: 0.046875rem;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-}
-
-.produtos-em-alta .topico-section .link-ver-todos {
-          font-size: 1.125rem;
-          font-family: var(--font_primaria);
-          font-weight: var(--font_weight_normal);
-          color: var(--primary);
-          letter-spacing: 0.046875rem;
-          text-decoration: none;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          cursor: pointer;
-          transition: 0.5s ease;
-}
-
-.produtos-em-alta .topico-section .link-ver-todos:hover {
-          color: var(--tertiary);
-}
