+++ src/pages/InformacoesPage/InformacoesPage.jsx
@@ -1,0 +1,14 @@
+import Header from "../../Components/Header/Header";
+import Footer from "../../Components/Footer/Footer";
+
+function Informacoes() {
+  return (
+    <>
+      <Header />
+      <h1>Informacoes</h1>
+      <Footer />
+    </>
+  );
+}
+
+export default Informacoes;
