--- src/pages/LoginPage.jsx
@@ -1,17 +1,0 @@
-import HeaderMenor from "../Components/HeaderMenor/HeaderMenor"
-import Footer from "../Components/Footer/Footer"
-import Main from "../Components/Main/Main"
-
-
-function Login() {
-
-  return (
-    <>
-      <HeaderMenor />
-      <Main />
-      <Footer />
-    </>
-  )
-}
-
-export default Login
