+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,0 +1,15 @@
+import HeaderMenor from "../../Components/HeaderMenor/HeaderMenor";
+import Footer from "../../Components/Footer/Footer";
+import Main from "../../Components/Main/Main";
+
+function Login() {
+  return (
+    <>
+      <HeaderMenor />
+      <Main />
+      <Footer />
+    </>
+  );
+}
+
+export default Login;
