+++ src/pages/PaginaDeSucessoPage/PaginaDeSucessoPage.jsx
@@ -1,0 +1,14 @@
+import Header from "../../Components/Header/Header";
+import Footer from "../../Components/Footer/Footer";
+
+function Sucesso() {
+  return (
+    <>
+      <Header />
+      <h1>Sucesso</h1>
+      <Footer />
+    </>
+  );
+}
+
+export default Sucesso;
