--- src/pages/PedidosPage.jsx
@@ -1,15 +1,0 @@
-import Header from "../Components/Header/Header"
-import Footer from "../Components/Footer/Footer"
-
-function Pedidos() {
-
-  return (
-    <>
-      <Header />
-      <h1>Pedidos</h1>
-      <Footer />
-    </>
-  )
-}
-
-export default Pedidos
