+++ src/pages/PedidosPage/PedidosPage.jsx
@@ -1,0 +1,14 @@
+import Header from "../../Components/Header/Header";
+import Footer from "../../Components/Footer/Footer";
+
+function Pedidos() {
+  return (
+    <>
+      <Header />
+      <h1>Pedidos</h1>
+      <Footer />
+    </>
+  );
+}
+
+export default Pedidos;
