+++ src/pages/ProductListingPage/ProductListingPage.css
@@ -1,0 +1,260 @@
+.section-product-list .topico-product-list {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  align-content: center;
+  padding: 2rem 5rem;
+}
+
+.section-product-list .topico-product-list h1 {
+  font-size: 1rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.75px;
+  display: flex;
+  justify-content: center;
+  margin: 0;
+}
+
+.section-product-list .topico-product-list span {
+  font-size: 1rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_normal);
+  color: var(--black);
+  letter-spacing: 0.75px;
+}
+
+.section-product-list .topico-product-list .dropdown-ordernar-por {
+  border: 1px var(--black) solid;
+  border-radius: 0.25rem;
+  width: 20.75rem;
+  height: 3.75rem;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.section-product-list .topico-product-list .dropdown-ordernar-por h1 {
+  font-size: 1rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.75px;
+}
+
+.section-product-list .topico-product-list .dropdown-ordernar-por span {
+  font-size: 1rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_normal);
+  color: var(--black);
+  letter-spacing: 0.75px;
+}
+
+.section-product-list .topico-product-list .dropdown-ordernar-por img {
+  width: 1rem;
+}
+
+.section-product-list .corpo-product-list {
+  padding: 2rem 5rem 0 5rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+}
+
+.section-product-list .corpo-product-list .checkbox-filtrar-por {
+  width: 21.38vw;
+  height: 82.28vh;
+  background-color: var(--white);
+  border-radius: 0.25rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.section-product-list .corpo-product-list .checkbox-filtrar-por h1 {
+  font-size: 1rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.75px;
+}
+
+.checkbox-filtrar-por-center {
+  width: 18vw;
+  height: 75vh;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: space-between;
+}
+
+.section-product-list
+  .corpo-product-list
+  .checkbox-filtrar-por
+  .barra-filtrar-por {
+  color: var(--black);
+  width: 100%;
+}
+
+.section-product-list .corpo-product-list .checkbox-filtrar-por .topico-marka {
+  display: flex;
+  flex-direction: column;
+}
+
+.section-product-list
+  .corpo-product-list
+  .checkbox-filtrar-por
+  .topico-marka
+  h1 {
+  font-size: 0.875rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.25px;
+}
+
+.section-product-list
+  .corpo-product-list
+  .checkbox-filtrar-por
+  .topico-categoria {
+  display: flex;
+  flex-direction: column;
+}
+
+.section-product-list
+  .corpo-product-list
+  .checkbox-filtrar-por
+  .topico-categoria
+  h1 {
+  font-size: 0.875rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.25px;
+}
+
+.section-product-list .corpo-product-list .checkbox-filtrar-por .topico-genero {
+  display: flex;
+  flex-direction: column;
+}
+
+.section-product-list
+  .corpo-product-list
+  .checkbox-filtrar-por
+  .topico-genero
+  h1 {
+  font-size: 0.875rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.25px;
+}
+
+.section-product-list .corpo-product-list .checkbox-filtrar-por .topico-estado {
+  display: flex;
+  flex-direction: column;
+}
+
+.section-product-list
+  .corpo-product-list
+  .checkbox-filtrar-por
+  .topico-estado
+  h1 {
+  font-size: 0.875rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--font_weight_bold);
+  color: var(--black);
+  letter-spacing: 0.25px;
+}
+
+.inputs-filtrar-por {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.input-filtrar-por {
+  width: 1.375rem;
+  height: 1.375rem;
+  margin-bottom: 0.5rem;
+  transform: scale(1);
+  transition: transform 0.2s ease-in-out;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.check:checked {
+  background-color: var(--primary) !important;
+  width: 1.375rem;
+  height: 1.375rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.radio:checked {
+  background-color: var(--white) !important;
+  width: 1.375rem;
+  height: 1.375rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.input-filtrar-por[type="checkbox"],
+.input-filtrar-por[type="radio"] {
+  -webkit-appearance: none; /* Safari e Chrome */
+  -moz-appearance: none; /* Firefox */
+  appearance: none;
+  background-color: var(--white);
+  border: 1px var(--black) solid;
+}
+
+.input-filtrar-por[type="checkbox"]:cheked {
+  background-color: var(--primary);
+  border: 1px var(--primary) solid;
+}
+
+.input-filtrar-por[type="radio"]:cheked {
+  background-color: var(--white) !important;
+  border: 1px var(--primary) solid;
+}
+
+.input-filtrar-por[type="checkbox"]:checked::before {
+  content: url(../assets/img/seta_ok.png);
+  display: flex;
+  width: 1rem;
+  height: 1rem;
+  justify-content: center;
+  align-items: center;
+}
+
+.input-filtrar-por[type="radio"]:checked::before {
+  content: url(../assets/img/elipse.svg);
+  display: flex;
+  width: 2rem;
+  height: 2rem;
+  border-radius: 50%;
+  justify-content: center;
+  align-items: center;
+  margin: 0;
+  padding: 0;
+}
+
+.input-filtrar-por[type="radio"] {
+  border-radius: 50%;
+}
+
+.label-filtrar-por {
+  font-size: 1rem;
+  font-family: var(--font_primaria);
+  font-weight: var(--normal);
+  color: var(--black);
+  letter-spacing: 0.25px;
+  margin-left: 0.5rem;
+  margin-bottom: 0.5rem;
+}
